View Controller code cleanup
